requirements.txt
selenium
python-dotenv
jsonlines